Update solution for Day 1 part 2
The second part of the puzzle can be found at [1]. The link will only
work if you completed part 1.

This was a bit of a painful one I have to admit. C does not bring as
convenient string to number conversion functions like more modern
programming languages. Because of this I had to implement my own
matching algorithm and of course I did not think of all cases at first
which resulted in me having to manually track down where my code did not
find the right values...
On the other hand, I wanted this to be a bit more challenging and
restricting myself to using only the standard C libraries was definitely
interesting.

[1]: https://adventofcode.com/2023/day/1